requirements.txt
PyQt5
pyqtgraph
pyusb
numpy
scipy
pyshortcuts
PyYAML